Basic client side board functionality working
Updated newBoard function so that it will now clear the board and reset
the cells array when the new game button is clicked. Updated the
checkWinningCombos function so that now it logs to the user whether
there was a win and if there was a draw. It also updates the variable
over to reflect whether the game is finished.