Updated messages
Updated text of message that displays when there is a draw, the message
in the get game stats modal, changed the button to say ok within the
game stats modal (from cancel) and updated the new game button text
to reset every time there is a successful sign in. continued to
change to reset after first game in a session though.